vasp.relax.Relax now works

+++ vasp/tests/runrelax.py
@@ -8,7 +8,7 @@ def test(path):
   from quantities import kbar, eV, angstrom
   from lada.crystal import Structure
   from lada.vasp import Vasp
-  from lada.vasp.relax import relax
+  from lada.vasp.relax import relax, Relax
 
     
     
@@ -24,9 +24,13 @@ def test(path):
   vasp.set_smearing   = "metal", 0.01
   vasp.relaxation = "volume"
   vasp.add_specie = "Si", "{0}/pseudos/Si".format(path)
-  directory = mkdtemp()
+  directory = '/tmp/test/relax' # mkdtemp()
   try: 
-    result = relax(vasp, structure, outdir=directory, comm={'n': 2, 'ppn': 1})
+    functional = Relax(copy=vasp)
+    assert abs(functional.ediff - 1e-5) < 1e-8
+    assert functional.precision == 'Accurate'
+    result = functional(structure, outdir=directory, comm={'n': 2, 'ppn': 1},
+                        relaxation="volume ionic cellshape")
     assert result.success
     def sortme(a): return int(a.split('/')[-1])
     dirs = sorted(glob(join(join(directory, '*'), '[0-9]')), key=sortme)
@@ -36,8 +40,9 @@ def test(path):
     assert result.stress.units == kbar and all(abs(result.stress) < 1e0)
     assert result.forces.units == eV/angstrom and all(abs(result.forces) < 1e-1)
     assert result.total_energy.units == eV and all(abs(result.total_energy + 10.668652*eV) < 1e-2)
+
   finally: 
-    rmtree(directory)
+    if directory != '/tmp/test/relax': rmtree(directory)
     pass
 
 if __name__ == "__main__":
